mlops-azure/api/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# ML Libraries
scikit-learn==1.3.2
numpy==1.24.3
joblib==1.3.2

# Azure SDKs
azure-cosmos==4.5.1
opencensus-ext-azure==1.1.13
applicationinsights==0.11.10

# Utilities
python-dotenv==1.0.0
